+++ branches/Free/ACG/Scenegraph/TriangleNode.hh
@@ -82,7 +82,7 @@ public:
 
   unsigned int availableDrawModes() const;
 
-  void boundingBox(Vec3f& _bbMin, Vec3f& _bbMax);
+  void boundingBox(Vec3d& _bbMin, Vec3d& _bbMax);
   void draw(GLState& _state, unsigned int _drawMode);
   void pick(GLState& _state, PickTarget _target);
